chore: rename getResultChartOptions
