+++ src/components/chart/ResultChart.tsx
@@ -9,7 +9,7 @@ import {
   Legend,
 } from "chart.js";
 import { Line } from "react-chartjs-2";
-import { getResultChartOptions } from "./resultChartOptions";
+import { getResultChartOptions } from "./getResultChartOptions";
 
 ChartJS.register(
   CategoryScale,
