+++ src/components/chart/getResultChartOptions.ts
